multiple rename

+++ multiple/multiple.go
@@ -20,9 +20,9 @@ var (
 )
 
 // 将一个reader共享给n个分身reader, 同一份输入复制成n分输入
-func CopiedReader(reader io.Reader, multiple uint32, done <-chan bool, filters ...Filter) (inputs []io.Reader) {
+func CopiedReader(reader io.Reader, multiple int, done <-chan bool, filters ...Filter) (inputs []io.Reader) {
 	var uplinkWriters []*io.PipeWriter
-	for i := uint32(0); i < multiple; i++ {
+	for i := 0; i < multiple; i++ {
 		r, w := io.Pipe()
 		inputs = append(inputs, r)
 		uplinkWriters = append(uplinkWriters, w)
@@ -66,9 +66,9 @@ func CopiedReader(reader io.Reader, multiple uint32, done <-chan bool, filters .
 }
 
 // 将一个writer分身成n个, 并将接收的输出合并成一个
-func MergedWriter(writer io.Writer, n uint32, done <-chan bool, filters ...Filter) (outputs []io.Writer) {
+func MergedWriter(writer io.Writer, n int, done <-chan bool, filters ...Filter) (outputs []io.Writer) {
 	var dwlinkReaders []*io.PipeReader
-	for i := uint32(0); i < n; i++ {
+	for i := 0; i < n; i++ {
 		r, w := io.Pipe()
 		outputs = append(outputs, w)
 		dwlinkReaders = append(dwlinkReaders, r)
